Move files around to make it easier to develop
